Move ScriptDefinition to script package
